GetCommentList

+++ models/movie.go
@@ -10,7 +10,7 @@ import (
  )
 
 type MovieData struct {
-	MovieId int64 `json:"movie_id"`
+	MovieId int `json:"movie_id"`
 	Name string `json:"name"`
 	ReleaseDate string `json:"release_date"`
 	OpeningCrawl string `json:"opening_crawl"`
